+++ srcs/utils/file_io_utils.hpp
@@ -5,6 +5,7 @@
 
 bool        is_file_exists(const std::string &path);
 bool        is_dir(const std::string &path);
+std::string read_dir_tostring(const std::string &file_path);
 std::string read_file_tostring(const std::string &file_path);
 bool        remove_file(const std::string &file_path);
 bool        is_accessible(const std::string &file_path, int mode);
